Add lombok, refactor
