+++ src/main/java/ru/math/tversu/studentapp/controller/SecurityController.java
@@ -1,13 +1,12 @@
 package ru.math.tversu.studentapp.controller;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
+@Log4j2
 public class SecurityController {
-	private Logger logger = LoggerFactory.getLogger(SecurityController.class);
 
 	@GetMapping(value = {"/", "/login"})
 	public String login() {
