+++ src/main/java/ru/math/tversu/studentapp/model/object/Lesson.java
@@ -1,29 +1,16 @@
 package ru.math.tversu.studentapp.model.object;
 
+import lombok.Data;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 
 @Entity
+@Data
 public class Lesson {
-    @Id
-    private Integer id;
-    @Column(name = "name")
-    private String name;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
+	@Id
+	private Integer id;
+	@Column(name = "name")
+	private String name;
 }
