+++ src/main/java/ru/math/tversu/studentapp/model/object/LessonTime.java
@@ -1,5 +1,7 @@
 package ru.math.tversu.studentapp.model.object;
 
+import lombok.Data;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -7,47 +9,16 @@ import javax.validation.constraints.Pattern;
 import java.time.DayOfWeek;
 
 @Entity
+@Data
 public class LessonTime {
-    @Id
-    private Integer id;
-    @Pattern(regexp = "^(([0-9])|([0-1][0-9])|([2][0-3]))(:(([0-9])|([0-5][0-9])))?$")
-    @Column(name = "start_time")
-    private String startTime;
-    @Pattern(regexp = "^(([0-9])|([0-1][0-9])|([2][0-3]))(:(([0-9])|([0-5][0-9])))?$")
-    @Column(name = "end_time")
-    private String endTime;
-    @Column(name = "weekday")
-    private DayOfWeek weekday;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getStartTime() {
-        return startTime;
-    }
-
-    public void setStartTime(String startTime) {
-        this.startTime = startTime;
-    }
-
-    public String getEndTime() {
-        return endTime;
-    }
-
-    public void setEndTime(String endTime) {
-        this.endTime = endTime;
-    }
-
-    public DayOfWeek getWeekday() {
-        return weekday;
-    }
-
-    public void setWeekday(DayOfWeek weekday) {
-        this.weekday = weekday;
-    }
+	@Id
+	private Integer id;
+	@Pattern(regexp = "^(([0-9])|([0-1][0-9])|([2][0-3]))(:(([0-9])|([0-5][0-9])))?$")
+	@Column(name = "start_time")
+	private String startTime;
+	@Pattern(regexp = "^(([0-9])|([0-1][0-9])|([2][0-3]))(:(([0-9])|([0-5][0-9])))?$")
+	@Column(name = "end_time")
+	private String endTime;
+	@Column(name = "weekday")
+	private DayOfWeek weekday;
 }
