+++ src/main/java/ru/math/tversu/studentapp/model/object/Room.java
@@ -1,29 +1,16 @@
 package ru.math.tversu.studentapp.model.object;
 
+import lombok.Data;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 
 @Entity
+@Data
 public class Room {
-    @Id
-    private Integer id;
-    @Column(name = "number")
-    private String number;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
+	@Id
+	private Integer id;
+	@Column(name = "number")
+	private String number;
 }
