+++ src/main/java/ru/math/tversu/studentapp/model/object/Schedule.java
@@ -1,20 +1,13 @@
 package ru.math.tversu.studentapp.model.object;
 
+import lombok.Data;
+
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.List;
 
 @Entity
+@Data
 public class Schedule {
-    @Id
-    private Integer id;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
+	@Id
+	private Integer id;
 }
