+++ src/main/java/ru/math/tversu/studentapp/model/user/Student.java
@@ -1,19 +1,13 @@
 package ru.math.tversu.studentapp.model.user;
 
+import lombok.Data;
+
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 
 @Entity
+@Data
 public class Student extends User {
-
-    @ManyToOne
-    private StudyGroup studyGroup;
-
-    public StudyGroup getGroup() {
-        return studyGroup;
-    }
-
-    public void setGroup(StudyGroup studyGroup) {
-        this.studyGroup = studyGroup;
-    }
+	@ManyToOne
+	private StudyGroup studyGroup;
 }
