+++ src/main/java/ru/math/tversu/studentapp/model/user/Teacher.java
@@ -1,19 +1,13 @@
 package ru.math.tversu.studentapp.model.user;
 
+import lombok.Data;
+
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 
 @Entity
+@Data
 public class Teacher extends User {
-
-    @ManyToOne
-    private EducatorTitle title;
-
-    public EducatorTitle getTitle() {
-        return title;
-    }
-
-    public void setTitle(EducatorTitle title) {
-        this.title = title;
-    }
+	@ManyToOne
+	private EducatorTitle title;
 }
