+++ src/main/java/ru/math/tversu/studentapp/repository/object/ScheduleItemRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.object.ScheduleItem;
 import java.util.List;
 
 public interface ScheduleItemRepository extends CrudRepository<ScheduleItem, Integer> {
-    List<ScheduleItem> findAll();
+	List<ScheduleItem> findAll();
 }
