+++ src/main/java/ru/math/tversu/studentapp/repository/object/ScheduleRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.object.Schedule;
 import java.util.List;
 
 public interface ScheduleRepository extends CrudRepository<Schedule, Integer> {
-    List<Schedule> findAll();
+	List<Schedule> findAll();
 }
