+++ src/main/java/ru/math/tversu/studentapp/repository/user/GroupRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.user.StudyGroup;
 import java.util.List;
 
 public interface GroupRepository extends CrudRepository<StudyGroup, Integer> {
-    List<StudyGroup> findAll();
+	List<StudyGroup> findAll();
 }
