+++ src/main/java/ru/math/tversu/studentapp/repository/user/StudentRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.user.Student;
 import java.util.List;
 
 public interface StudentRepository extends CrudRepository<Student, Integer> {
-    List<Student> findAll();
+	List<Student> findAll();
 }
