+++ src/main/java/ru/math/tversu/studentapp/repository/user/TeacherRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.user.Teacher;
 import java.util.List;
 
 public interface TeacherRepository extends CrudRepository<Teacher, Integer> {
-    List<Teacher> findAll();
+	List<Teacher> findAll();
 }
