+++ src/main/java/ru/math/tversu/studentapp/repository/user/TitleRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.user.EducatorTitle;
 import java.util.List;
 
 public interface TitleRepository extends CrudRepository<EducatorTitle, Integer> {
-    List<EducatorTitle> findAll();
+	List<EducatorTitle> findAll();
 }
