+++ src/main/java/ru/math/tversu/studentapp/repository/user/UserRepository.java
@@ -6,5 +6,5 @@ import ru.math.tversu.studentapp.model.user.User;
 import java.util.List;
 
 public interface UserRepository extends CrudRepository<User, Integer> {
-    List<User> findAll();
+	List<User> findAll();
 }
