+++ src/main/resources/templates/login.html
@@ -7,9 +7,9 @@
     <title>App</title>
 </head>
 <body>
-    <h3 th:inline="text">Welcome [[${#httpServletRequest.remoteUser}]]</h3>
-    <form th:action="@{/logout}" method="post">
-        <input type="submit" value="Logout" />
-    </form>
+<h3 th:inline="text">Welcome [[${#httpServletRequest.remoteUser}]]</h3>
+<form th:action="@{/logout}" method="post">
+    <input type="submit" value="Logout"/>
+</form>
 </body>
 </html>
